+++ src/Laravel/workbench/app/Models/BookVariant.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the API Platform project.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Workbench\App\Models;
+
+/**
+ * Not a resource itself, but an instance of one: IRI generation promotes it to Book.
+ */
+class BookVariant extends Book
+{
+    protected $table = 'books';
+}